fix: renamed the file, defined props explicitly and used our Button Component for closing Alert
